feat: update time calculation methods to return Int and adjust service update parameters in MapFragment

+++ app/src/main/java/gorda/driver/maps/Map.kt
@@ -32,11 +32,11 @@ object Map {
         else String.format("%.2f", distance) + "m"
     }
 
-    fun calculateTime(distance: Float): Float {
-        return (distance / 5)
+    fun calculateTime(distance: Float): Int {
+        return (distance / 5).toInt()
     }
 
-    fun getTimeString(time: Float): String {
+    fun getTimeString(time: Int): String {
         return if (time < 60) {
             "1m"
         } else {
